refactor(popmenu): category-refactor
